Fix the positioning of the New Admin screenshot.
The New Admin screenshot wasn't being positioned correctly. This was due
to the positioning logic, which assumed a screenshot would fit on the
page, factoring in the offsets on the side.

Now if the image is just too big, we decide to use the page width as the
max width, centering the image properly.